Move API models to their own class
Out of the map cache into their own files
